refactor: remove functions and events duplicate code
