+++ web3/types.py
@@ -36,14 +36,9 @@ from web3._utils.compat import (
 )
 
 if TYPE_CHECKING:
-    from web3.contract.async_contract import (  # noqa: F401
-        AsyncContractEvent,
-    )
     from web3.contract.base_contract import (
+        BaseContractEvent,
         BaseContractFunction,
-    )
-    from web3.contract.contract import (  # noqa: F401
-        ContractEvent,
     )
     from web3.main import (  # noqa: F401
         AsyncWeb3,
@@ -583,7 +578,7 @@ class GethWallet(TypedDict):
 # Contract types
 
 TContractFn = TypeVar("TContractFn", bound="BaseContractFunction")
-TContractEvent = TypeVar("TContractEvent", "ContractEvent", "AsyncContractEvent")
+TContractEvent = TypeVar("TContractEvent", bound="BaseContractEvent")
 
 
 # Tracing types
